requirements.txt
matplotlib==3.3.4
numpy==1.19.2
python==3.6.13
pytorch==1.10.1
gensim==4.2.0
karateclub==1.3.0
networkx==2.4
node2vec==0.4.3
pandas==1.1.5
python-louvain==0.16
scikit-learn==0.24.2
scipy==1.5.4
stellargraph==1.2.1
torch-geometric==2.0.3

